requirements.txt
# Core inference + detection
torch>=2.1.0
ultralytics>=8.2.0

# Synthetic data generation
blenderproc>=2.6.0

# VLM API client
httpx>=0.27.0

# Configuration
pyyaml>=6.0

# Numerical computing
numpy>=1.26.0

# Evaluation (COCO standard mAP)
pycocotools>=2.0.7

# REST API server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9    # Required for FastAPI file upload

# Pydantic v2 (FastAPI dependency, listed explicitly for Jetson pin)
pydantic>=2.7.0

# Image handling (EWC consolidation DataLoader)
Pillow>=10.3.0
